Problem 9 - Brute Force
There are 3 nested loops that each go up to 1000. This solution can be improved, but it works!

+++ Project Euler/Project Euler/Problem8.cpp
@@ -2,7 +2,7 @@
 #include <fstream>
 #include <vector>
 
-unsigned long long Problem8::Solve(std::string& _fileName, int _digitCount)
+unsigned long long Problem8::Solve(std::string& _fileName, unsigned long long _digitCount)
 {
 	//do the maths
 	std::vector<int> digits;
